shared/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
httpx==0.27.2
orjson==3.10.7
aiokafka==0.11.0
python-jose[cryptography]==3.3.0
